refactor: moved logic for budget month spend chart to a seperate utils folder. feat: added a filter to show limited categories
